[#15] UI : 해당 날짜 일기 쓰기(수정)
