+++ app/src/main/res/layout/fragment_write_diary.xml
@@ -1,0 +1,63 @@
+<?xml version="1.0" encoding="utf-8"?>
+<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    tools:context=".WriteDiaryFragment">
+    <androidx.constraintlayout.widget.ConstraintLayout
+        android:layout_width="match_parent"
+        android:layout_height="match_parent">
+        <ImageView
+            android:id="@+id/write_back_btn"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_marginEnd="16dp"
+            android:src="@drawable/ic_close"
+            app:layout_constraintBottom_toBottomOf="@id/write_date_tv"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintTop_toTopOf="@id/write_date_tv" />
+
+        <TextView
+            android:id="@+id/write_date_tv"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_marginTop="38dp"
+            android:fontFamily="@font/pretendard"
+            android:text="12월 5일 목요일"
+            android:textSize="20sp"
+            android:textStyle="bold"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toTopOf="parent" />
+
+        <TextView
+            android:id="@+id/diary_editText_tv"
+            android:layout_width="292dp"
+            android:layout_height="352dp"
+            android:layout_marginTop="40dp"
+            android:paddingHorizontal="16dp"
+            android:paddingBottom="77dp"
+            android:background="@drawable/ic_diary_edit_box"
+            android:text=""
+            android:textSize="14sp"
+            android:elevation="4dp"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toBottomOf="@id/write_date_tv" />
+
+        <ImageView
+            android:id="@+id/write_finish_btn"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_marginBottom="15dp"
+            android:src="@drawable/btn_completed"
+            android:elevation ="8dp"
+            app:layout_constraintBottom_toBottomOf="@id/diary_editText_tv"
+            app:layout_constraintEnd_toEndOf="@id/diary_editText_tv"
+            app:layout_constraintStart_toStartOf="@id/diary_editText_tv" />
+
+
+    </androidx.constraintlayout.widget.ConstraintLayout>
+
+</FrameLayout>
